Refactor Layout component
